requirements.txt
streamlit==1.28.0
google-generativeai==0.3.0
pypdf==3.17.0
python-dotenv==0.19.0